# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "NewsBot collector",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/config/sources/by/source": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Lists the top 50 records based on the name given. Example: reddit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source Name",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/by/sourceAndName": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Returns a single entity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dadjokes",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "reddit",
                        "name": "source",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/new/reddit": {
            "post": {
                "tags": [
                    "Config",
                    "Source",
                    "Reddit"
                ],
                "summary": "Creates a new reddit source to monitor.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/new/twitch": {
            "post": {
                "tags": [
                    "Config",
                    "Source",
                    "Twitch"
                ],
                "summary": "Creates a new twitch source to monitor.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/new/youtube": {
            "post": {
                "tags": [
                    "Config",
                    "Source",
                    "YouTube"
                ],
                "summary": "Creates a new youtube source to monitor.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Returns a single entity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Deletes a record by ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/{id}/disable": {
            "post": {
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Disables a source from processing.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/config/sources/{id}/enable": {
            "post": {
                "tags": [
                    "Config",
                    "Source"
                ],
                "summary": "Enables a source to continue processing.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/subscriptions/byDiscordId": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Subscription"
                ],
                "summary": "Returns the top 100 entries from the queue to be processed.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/subscriptions/bySourceId": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config",
                    "Subscription"
                ],
                "summary": "Returns the top 100 entries from the queue to be processed.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/subscriptions/discord/webhook/delete": {
            "post": {
                "tags": [
                    "Config",
                    "Source",
                    "Discord",
                    "Subscription"
                ],
                "summary": "Removes a Discord WebHook Subscription based on the Subscription ID.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "Id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/subscriptions/new/discordwebhook": {
            "post": {
                "tags": [
                    "Config",
                    "Source",
                    "Discord",
                    "Subscription"
                ],
                "summary": "Creates a new subscription to link a post from a Source to a DiscordWebHook.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "discordWebHookId",
                        "name": "discordWebHookId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "sourceId",
                        "name": "sourceId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}